+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_help.py
@@ -505,7 +505,98 @@ helps['network application-gateway redirect-config update'] = """
             az network application-gateway redirect-config update -g MyResourceGroup --gateway-name MyAppGateway \\
                 -n MyRedirectConfig --type Permanent --target-listener MyNewBackendListener
 """
+# endregion
 
+# region Application Gateway rewrite rules
+helps['network application-gateway rewrite-rule'] = """
+    short-summary: Manage rewrite rules of an application gateway.
+    type: group
+"""
+
+helps['network application-gateway rewrite-rule create'] = """
+    short-summary: Create a rewrite rule.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule delete'] = """
+    short-summary: Delete a rewrite rule.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule list'] = """
+    short-summary: List rewrite rules.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule show'] = """
+    short-summary: Get the details of a rewrite rule.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule update'] = """
+    short-summary: Update a rewrite rule.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule set'] = """
+    short-summary: Manage rewrite rule sets of an application gateway.
+    type: group
+"""
+
+helps['network application-gateway rewrite-rule set create'] = """
+    short-summary: Create a rewrite rule set.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule set delete'] = """
+    short-summary: Delete a rewrite rule set.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule set list'] = """
+    short-summary: List rewrite rule sets.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule set show'] = """
+    short-summary: Get the details of a rewrite rule set.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule set update'] = """
+    short-summary: Update a rewrite rule set.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule condition'] = """
+    short-summary: Manage rewrite rule conditions of an application gateway.
+    type: group
+"""
+
+helps['network application-gateway rewrite-rule condition create'] = """
+    short-summary: Create a rewrite rule condition.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule condition delete'] = """
+    short-summary: Delete a rewrite rule condition.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule condition list'] = """
+    short-summary: List rewrite rule conditions.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule condition show'] = """
+    short-summary: Get the details of a rewrite rule condition.
+    type: command
+"""
+
+helps['network application-gateway rewrite-rule condition update'] = """
+    short-summary: Update a rewrite rule condition.
+    type: command
+"""
 # endregion
 
 # region Application Gateway rules