+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,8 +2,10 @@
 
 Release History
 ===============
-
+0.2.16
+++++++
 * webapp: az webapp up was not handling running from empty directory or unknown code scenario correctly
+* webapp, functionapp: az webapp config ssl bind command was failing for slots
 
 0.2.15
 ++++++