+++ src/command_modules/azure-cli-interactive/HISTORY.rst
@@ -2,7 +2,8 @@
 
 Release History
 ===============
-
+0.4.2
++++++
 * Fix incompatibility with Interactive extension installed through azdev.
 
 0.4.1