+++ src/command_modules/azure-cli-role/HISTORY.rst
@@ -2,6 +2,8 @@
 
 Release History
 ===============
+2.4.3
++++++
 * `ad app/sp credential reset`: support --end-date
 * `ad app permission add`: support to add permissions on existing api
 * `ad app permission list`: fix a bad error when there is no permissions